+++ packages/music/src/db/database.spec.ts
@@ -1,9 +1,7 @@
-import { loadEnv } from "../env";
 import { connect, disconnect } from "./database";
 
 describe("connect", () => {
   it("connection", async () => {
-    loadEnv();
     await connect();
     await disconnect();
   } );
