--- packages/music/src/db/models/music/music.delete.ts
@@ -1,6 +1,0 @@
-import MusicModel from "./music.model";
-
-// eslint-disable-next-line import/prefer-default-export
-export async function deleteAll() {
-  await MusicModel.deleteMany();
-}
