--- packages/music/src/db/models/music/music.mock.ts
@@ -1,9 +1,0 @@
-import { createFromPathAndSave, deleteAll } from ".";
-
-export async function initializeMock() {
-  await createFromPathAndSave("dk.mp3");
-}
-
-export async function clearMock() {
-  await deleteAll();
-}
