+++ packages/music/src/env.ts
@@ -3,7 +3,7 @@ import path from "path";
 let alreadyLoaded = false;
 
 // eslint-disable-next-line import/prefer-default-export
-export function loadEnv() {
+export function assertEnv() {
   if (alreadyLoaded)
     return;
 
@@ -22,7 +22,7 @@ export function checkEnvVar(name: string) {
 }
 
 export function getFullPath(relativePath: string): string {
-  loadEnv();
+  assertEnv();
   const MEDIA_PATH = <string>process.env.MEDIA_PATH;
 
   return path.join(MEDIA_PATH, relativePath);
