+++ packages/music/src/files/files.hash.ts
@@ -1,5 +1,4 @@
 /* eslint-disable import/prefer-default-export */
-/* eslint-disable no-param-reassign */
 import sha256File from "sha256-file";
 
 export function calcHashFromFile(filepath: string) {
