+++ packages/music/src/files/files.spec.ts
@@ -1,3 +1,4 @@
+import { join } from "node:path";
 import { calcHashFromFile, findFiles, FindOptions, fixHashFile } from ".";
 
 describe("getHashFromFile", () => {
@@ -146,11 +147,11 @@ describe("findFilesByExtensionRecursive", () => {
 
 describe("findFiles", () => {
   it("unique hash", () => {
-    const expected = ["tests/files/dk.mp3"];
+    const expected = [join(process.env.MEDIA_PATH ?? "", "dk.mp3")];
     const hash = "eacf40b68de85b759524e3bd0bea1b4393360f682db3a7f3ec25ff46b1d01872";
     const options: FindOptions = {
       fileHash: hash,
-      folder: "tests/files",
+      folder: process.env.MEDIA_PATH,
       recursive: false,
     };
     const actual = findFiles(options);
@@ -163,7 +164,7 @@ describe("findFiles", () => {
     const hash = "54ca5061257adafcedee8523e4f8cc3f0347ab9143cddb0fd9b4997498e34ce2";
     const options: FindOptions = {
       fileHash: hash,
-      folder: "tests/files",
+      folder: process.env.MEDIA_PATH,
       recursive: false,
     };
     const actual = findFiles(options);
@@ -176,7 +177,7 @@ describe("findFiles", () => {
     const hash = "1234";
     const options: FindOptions = {
       fileHash: hash,
-      folder: "tests/files",
+      folder: process.env.MEDIA_PATH,
       recursive: false,
     };
     const actual = findFiles(options);
@@ -191,7 +192,7 @@ describe("findFilesRecursive", () => {
     const hash = "eacf40b68de85b759524e3bd0bea1b4393360f682db3a7f3ec25ff46b1d01872";
     const options = {
       fileHash: hash,
-      folder: "tests/files",
+      folder: process.env.MEDIA_PATH,
     };
     const actual = findFiles(options);
 
@@ -203,7 +204,7 @@ describe("findFilesRecursive", () => {
     const hash = "54ca5061257adafcedee8523e4f8cc3f0347ab9143cddb0fd9b4997498e34ce2";
     const options = {
       fileHash: hash,
-      folder: "tests/files",
+      folder: process.env.MEDIA_PATH,
     };
     const actual = findFiles(options);
 
@@ -215,7 +216,7 @@ describe("findFilesRecursive", () => {
     const hash = "1234";
     const options = {
       fileHash: hash,
-      folder: "tests/files",
+      folder: process.env.MEDIA_PATH,
     };
     const actual = findFiles(options);
