+++ packages/music/src/models/musics/index.ts
@@ -1,0 +1,1 @@
+// TODO: Aquí poner los modelos al margen del ODM, los que irán en shared
