+++ packages/music/src/modules/musics/controllers/index.ts
@@ -1,0 +1,3 @@
+export {
+  default as Controller,
+} from "./Controller";
