+++ packages/music/src/modules/musics/controllers/routes.get-all.spec.ts
@@ -1,6 +1,6 @@
-import { createFromPath, Music } from "../../db/models/music";
-import { TestingApp1 } from "../TestingApps";
-import { generateView } from "./routes.get-all";
+import { MusicDocOdm } from "#modules/musics";
+import { TestingApp1 } from "../../../routes/tests/TestingApps";
+import { generateView } from "../repositories/Repository";
 
 const app = new TestingApp1();
 
@@ -12,7 +12,7 @@ afterAll(async () => {
   await app.kill();
 } );
 it("generateView", async () => {
-  const musics: Music[] = [await createFromPath("dk.mp3")];
+  const musics: MusicDocOdm[] = [await app.getMusicRepository().createFromPath("dk.mp3")];
   const actual = generateView(musics);
   const expectedText = "<ul><li><a href='/raw/dk'>dk</li></ul>";
