+++ packages/music/src/modules/musics/controllers/routes.get-random.spec.ts
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-extraneous-dependencies */
 // eslint-disable-next-line import/no-extraneous-dependencies
 import request from "supertest";
-import App from "../app";
-import { TestingApp1 } from "../TestingApps";
+import App from "../../../routes/app";
+import { TestingApp1 } from "../../../routes/tests/TestingApps";
 
 const app: App = new TestingApp1();
