+++ packages/music/src/modules/musics/index.ts
@@ -1,0 +1,7 @@
+export {
+  DocOdm as MusicDocOdm, ModelOdm as MusicModelOdm, Repository as MusicRepository, SchemaOdm as MusicSchemaOdm,
+} from "./repositories";
+
+export {
+  Controller as MusicController,
+} from "./controllers";
