+++ packages/music/src/modules/musics/repositories/index.ts
@@ -1,0 +1,9 @@
+export {
+  DocOdm,
+  ModelOdm,
+  SchemaOdm,
+} from "./odm";
+
+export {
+  default as Repository,
+} from "./Repository";
