+++ packages/music/src/modules/musics/repositories/tests/music.spec.ts
@@ -1,8 +1,10 @@
-import { findByHash, findByUrl } from ".";
-import App from "../../../routes/app";
-import { TestingApp1 } from "../../../routes/TestingApps";
+import { MusicRepository } from "#modules/musics";
+import App from "../../../../routes/app";
+import { TestingApp1 } from "../../../../routes/tests/TestingApps";
 
 describe("all tests", () => {
+  const repository = new MusicRepository();
+
   describe("getByHash", () => {
     const app: App = new TestingApp1();
 
@@ -15,14 +17,14 @@ describe("all tests", () => {
     } );
 
     it("not found", async () => {
-      const actual = await findByHash("asd");
+      const actual = await repository.findByHash("asd");
 
       expect(actual).toBeNull();
     } );
 
     it("found", async () => {
       const HASH = "eacf40b68de85b759524e3bd0bea1b4393360f682db3a7f3ec25ff46b1d01872";
-      const actual = await findByHash(HASH);
+      const actual = await repository.findByHash(HASH);
 
       expect(actual).not.toBeNull();
       expect(actual?.hash).toBe(HASH);
@@ -42,7 +44,7 @@ describe("all tests", () => {
 
     it("found", async () => {
       const URL = "dk";
-      const actual = await findByUrl(URL);
+      const actual = await repository.findByUrl(URL);
 
       expect(actual).not.toBeNull();
       expect(actual?.url).toBe(URL);
