--- packages/music/src/routes/api/routes.api.config.ts
@@ -1,21 +1,0 @@
-export const API = "/api";
-
-export const GET = `${API}/get`;
-
-export const CREATE = `${API}/create`;
-
-export const UPDATE = `${API}/update`;
-
-export const FIX = `${UPDATE}/fix`;
-
-export const ROUTE_RAW = `${GET}/raw`;
-
-export const ROUTE_GET_ALL = `${GET}/all`;
-
-export const ROUTE_GET_RANDOM = `${GET}/random`;
-
-export const ROUTE_CREATE_YT = `${CREATE}/yt`;
-
-export const ROUTE_FIX_ALL = `${FIX}/all`;
-
-export const ROUTE_FIX_ONE = `${FIX}/one`;
